Empty state for support tickets and ticket overview
